% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docket.R
\name{comment_counts}
\alias{comment_counts}
\title{Number of comments associated with a document}
\usage{
comment_counts(doc_id)
}
\arguments{
\item{doc_id}{The API identifier for the document, eg, `'090000648320bc9e'`}
}
\value{
Value in the `meta:totalElements` field
}
\description{
Queries the API and reports the associated comment count.
}
