% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_text.R
\name{extract_text}
\alias{extract_text}
\title{Extract text from a PDF or DOCX to a TXT file}
\usage{
extract_text(source, target, force = FALSE)
}
\arguments{
\item{source}{Source file}

\item{target}{Target TXT file}

\item{force}{Extract text even if target file already exists}
}
\value{
Extracted text (invisible)
}
\description{
Extract text from a PDF or DOCX to a TXT file
}
