% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_text.R
\name{extract_text_pdf}
\alias{extract_text_pdf}
\title{Extract text from a PDF file}
\usage{
extract_text_pdf(source, ocr_threshold = 50)
}
\arguments{
\item{source}{Source file to extract test from}

\item{ocr_threshold}{If text length in characters is less than this, run it through OCR}
}
\description{
Extract text from a PDF file
}
