% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docket.R
\name{get_blocks}
\alias{get_blocks}
\title{Get metadata for all comments associated with a document}
\usage{
get_blocks(doc_id, total_comments = NULL, verbose = FALSE)
}
\arguments{
\item{doc_id}{API document identifier, eg, `'090000648320bc9e'`}

\item{total_comments}{Total number of comments associated with the document.  If `NULL`, API will be queried to get this number.}

\item{verbose}{Passed down to `get_block`}
}
\value{
A data frame containing the comment metadata
}
\description{
Admittedly not the most perspicuous name
}
