% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docket.R
\name{parser}
\alias{parser}
\title{Parse scraped comment files}
\usage{
parser(file, verbose = FALSE)
}
\arguments{
\item{file}{The scraped comment JSON file}

\item{verbose}{Print the path to the file being parsed}
}
\value{
A data frame with the comment metadata and text.  If the comment has any attachments, metadata for these are given in the nested data frame `attachments`.
}
\description{
Parse scraped comment files
}
