% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape}
\alias{scrape}
\title{Scrape a file using `curl`}
\usage{
scrape(url, dl_target, force = FALSE, delay = 0, verbose = FALSE, ...)
}
\arguments{
\item{url}{URL of the file to download}

\item{dl_target}{Path to the target file}

\item{force}{Force redownload?}

\item{delay}{Delay in seconds after completing download}

\item{verbose}{Display the URL being downloaded}

\item{...}{Other parameters passed to `curl_download`}
}
\value{
Value of `dl_target` (invisibly)
}
\description{
A wrapper around `curl::curl_download` that checks whether the target file already exists.
}
