% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docket.R
\name{scrape_comment}
\alias{scrape_comment}
\title{Scrape the data for a comment}
\usage{
scrape_comment(
  comment_id,
  scrape_folder = file.path(data_folder, str_c(prefix, "comments")),
  ...
)
}
\arguments{
\item{comment_id}{ID of the comment to scrape}

\item{scrape_folder}{Folder where the comment should be saved}

\item{...}{Other options, passed to `scrape()`}
}
\value{
Path to the scraped comment (invisible)
}
\description{
Scrape the data for a comment
}
\note{
Scraping comments is strongly constrainted by the limit of 1000 queries/hour.  `scrape()` checks whether a file has already been downloaded, facilitating stopping and restarting this function.
}
