% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docket.R
\name{wrangle_date}
\alias{wrangle_date}
\title{Subtract 5 hours and add 1 second to a date-time character string}
\usage{
wrangle_date(json_date)
}
\arguments{
\item{json_date}{The date-time character string}
}
\value{
A date-time character string, formatted as required by the API
}
\description{
This function is used by `get_blocks` to page through the blocks of 5000 comments:  get the date-time for the last comment of the previous block, add 1 second, and use that as the threshold for the next block.  However, the API reports date-times in the format `YYYY-MM-DDTHH:mm:ssZ`, using UTC, while the query language only accepts the format `YYYY-MM-DD HH:mm:ss` using Eastern time.
}
\note{
API documentation doesn't specify whether this is Eastern Standard Time or whether it might be daylight savings (depending on when the comment was submitted, etc.)
}
